+++ IcVibracoes.Common/Classes/MechanicalProperty.cs
@@ -5,7 +5,7 @@ namespace IcVibracoes.Common.Classes
     /// <summary>
     /// It contains the mechanical properties to be used in Rigid Body analysis.
     /// </summary>
-    public class MechanicalProperties
+    public class MechanicalProperty
     {
         /// <summary>
         /// The mass of the analyzed object.